refactor : move UserId package QUZ-81
