+++ quiz-service/quiz-application/app-api/src/main/kotlin/com/grepp/quizy/quiz/api/comment/dto/CreateCommentRequest.kt
@@ -3,7 +3,7 @@ package com.grepp.quizy.quiz.api.comment.dto
 import com.grepp.quizy.quiz.domain.comment.CommentContent
 import com.grepp.quizy.quiz.domain.comment.CommentId
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 class CreateCommentRequest(
         val quizId: Long,
