+++ quiz-service/quiz-application/app-api/src/main/kotlin/com/grepp/quizy/quiz/api/like/LikeApi.kt
@@ -2,11 +2,11 @@ package com.grepp.quizy.quiz.api.like
 
 import com.grepp.quizy.common.api.ApiResponse
 import com.grepp.quizy.quiz.api.like.dto.LikeRequest
-import com.grepp.quizy.quiz.api.like.dto.LikeStatusResponse
 import com.grepp.quizy.quiz.domain.like.LikeService
-import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/api/quiz/likes")
