+++ quiz-service/quiz-application/app-api/src/main/kotlin/com/grepp/quizy/quiz/api/like/dto/LikeRequest.kt
@@ -2,7 +2,7 @@ package com.grepp.quizy.quiz.api.like.dto
 
 import com.grepp.quizy.quiz.domain.like.Like
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 data class LikeRequest(val quizId: Long, val userId: Long) {
