+++ quiz-service/quiz-application/app-api/src/main/kotlin/com/grepp/quizy/quiz/api/like/dto/LikeStatusCheckRequest.kt
@@ -1,7 +1,7 @@
 package com.grepp.quizy.quiz.api.like.dto
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 data class LikeStatusCheckRequest(
         val userId: Long,
