+++ quiz-service/quiz-application/app-api/src/main/kotlin/com/grepp/quizy/quiz/api/useranswer/dto/UserAnswerRequest.kt
@@ -2,7 +2,7 @@ package com.grepp.quizy.quiz.api.useranswer.dto
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
 import com.grepp.quizy.quiz.domain.useranswer.UserAnswerId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 data class UserAnswerRequest(
         val userId: Long,
