+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/comment/CommentCreateUseCase.kt
@@ -1,7 +1,7 @@
 package com.grepp.quizy.quiz.domain.comment
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 interface CommentCreateUseCase {
     fun createComment(
