+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/comment/CommentUpdater.kt
@@ -1,6 +1,6 @@
 package com.grepp.quizy.quiz.domain.comment
 
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 import org.springframework.stereotype.Component
 
 @Component
