+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/like/Like.kt
@@ -1,6 +1,6 @@
 package com.grepp.quizy.quiz.domain.like
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 data class Like(val likerId: UserId, val quizId: QuizId) {}
