+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/like/LikeManager.kt
@@ -2,7 +2,7 @@ package com.grepp.quizy.quiz.domain.like
 
 import com.grepp.quizy.quiz.domain.quiz.QuizCache
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 import org.springframework.stereotype.Component
 
 @Component
