+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/user/UserVO.kt
@@ -1,3 +1,3 @@
 package com.grepp.quizy.quiz.domain.user
 
-@JvmInline value class UserId(val id: Long)
+@JvmInline value class UserId(val value: Long)
