+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/useranswer/UserAnswerId.kt
@@ -1,5 +1,6 @@
 package com.grepp.quizy.quiz.domain.useranswer
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 data class UserAnswerId(val userId: UserId, val quizId: QuizId)
