+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/useranswer/UserAnswerReader.kt
@@ -1,6 +1,7 @@
 package com.grepp.quizy.quiz.domain.useranswer
 
 import com.grepp.quizy.quiz.domain.quiz.QuizId
+import com.grepp.quizy.quiz.domain.user.UserId
 import org.springframework.stereotype.Component
 
 @Component
