--- quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/useranswer/UserAnswerVO.kt
@@ -1,3 +1,0 @@
-package com.grepp.quizy.quiz.domain.useranswer
-
-@JvmInline value class UserId(val value: Long)
