+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/like/entity/LikeEntity.kt
@@ -3,7 +3,7 @@ package com.grepp.quizy.quiz.infra.like.entity
 import com.grepp.quizy.jpa.BaseTimeEntity
 import com.grepp.quizy.quiz.domain.like.Like
 import com.grepp.quizy.quiz.domain.quiz.QuizId
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 import jakarta.persistence.EmbeddedId
 import jakarta.persistence.Entity
 import jakarta.persistence.Table
