+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/quiz/entity/ABTestEntity.kt
@@ -2,7 +2,7 @@ package com.grepp.quizy.quiz.infra.quiz.entity
 
 import com.grepp.quizy.common.dto.DateTime
 import com.grepp.quizy.quiz.domain.quiz.*
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 import jakarta.persistence.DiscriminatorValue
 import jakarta.persistence.Entity
