+++ quiz-service/quiz-infra/src/main/kotlin/com/grepp/quizy/quiz/infra/quizread/document/QuizDomainFactory.kt
@@ -2,7 +2,7 @@ package com.grepp.quizy.quiz.infra.quizread.document
 
 import com.grepp.quizy.common.dto.DateTime
 import com.grepp.quizy.quiz.domain.quiz.*
-import com.grepp.quizy.quiz.domain.useranswer.UserId
+import com.grepp.quizy.quiz.domain.user.UserId
 
 class QuizDomainFactory {
     companion object {
